+++ Libplanet.Tests/Action/AccountStateDeltaExtensions.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
+using System.Numerics;
 using Bencodex.Types;
 using Libplanet.Action;
 
@@ -12,12 +13,24 @@ namespace Libplanet.Tests.Action
             this IAccountStateDelta delta
         )
         {
-            return delta.UpdatedAddresses.Select(address =>
+            return delta.StateUpdatedAddresses.Select(address =>
                 new KeyValuePair<Address, IValue>(
                     address,
                     delta.GetState(address)
                 )
             ).ToImmutableDictionary();
         }
+
+        public static IImmutableDictionary<(Address, Currency), BigInteger> GetUpdatedBalances(
+            this IAccountStateDelta delta
+        ) =>
+            delta.UpdatedFungibleAssets.SelectMany(kv =>
+                kv.Value.Select(currency =>
+                    new KeyValuePair<(Address, Currency), BigInteger>(
+                        (kv.Key, currency),
+                        delta.GetBalance(kv.Key, currency)
+                    )
+                )
+            ).ToImmutableDictionary();
     }
 }